refactor: move feature hooks out of containers
