+++ src/features/card/containers/CardFormContainer.tsx
@@ -6,7 +6,7 @@ import * as C from "@src/constant";
 import * as selector from "@src/selector";
 import { useActions } from "@src/shared/hooks/useActions";
 import { CardFormTemplate } from "@src/features/card/components/templates/CardFormTemplate";
-import { useCardFormState } from "@src/features/card/containers/useCardFormState";
+import { useCardFormState } from "@src/features/card/hooks/useCardFormState";
 
 export const CardFormContainer: React.FC = () => {
   const params = useParams();
