+++ src/features/settings/containers/ConfigContainer.tsx
@@ -3,7 +3,7 @@ import { useSelector } from "react-redux";
 import { useKey } from "react-use";
 
 import { ConfigFormTemplate } from "@src/features/settings/components/templates/ConfigFormTemplate";
-import { useConfigFormState } from "@src/features/settings/containers/useConfigFormState";
+import { useConfigFormState } from "@src/features/settings/hooks/useConfigFormState";
 import * as selector from "@src/selector";
 import { useActions } from "@src/shared/hooks/useActions";
