+++ src/features/study/containers/DeckSwiperContainer.tsx
@@ -11,9 +11,9 @@ import { CardOverlay } from "@src/features/card/components/CardOverlay";
 import { FrontText } from "@src/features/card/components/FrontText";
 import { DeckSwiperTemplate } from "@src/features/study/components/templates/DeckSwiperTemplate";
 import type { SwipeButtonListProps } from "@src/features/study/components/SwipeButtonList";
-import { getLegacyStudyCandidate, useLegacyStudySession } from "@src/features/study/containers/useLegacyStudySession";
-import { useStudyActions } from "@src/features/study/containers/useStudyActions";
-import { useStudyControllerState } from "@src/features/study/containers/useStudyControllerState";
+import { getLegacyStudyCandidate, useLegacyStudySession } from "@src/features/study/hooks/useLegacyStudySession";
+import { useStudyActions } from "@src/features/study/hooks/useStudyActions";
+import { useStudyControllerState } from "@src/features/study/hooks/useStudyControllerState";
 import { studyStore, useStudyStore } from "@src/features/study/state/studyStore";
 import { useActions } from "@src/shared/hooks/useActions";
