+++ src/features/study/hooks/useLegacyStudySession.spec.tsx
@@ -4,7 +4,7 @@ import { cleanup, render, waitFor } from "@testing-library/react";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
 import * as type from "@src/action/type";
-import { useLegacyStudySession } from "@src/features/study/containers/useLegacyStudySession";
+import { useLegacyStudySession } from "@src/features/study/hooks/useLegacyStudySession";
 import { type LegacyStudyCandidate, studyStore } from "@src/features/study/state/studyStore";
 
 const mocks = vi.hoisted(() => ({
